safari-534.53-branch/LayoutTests/fast/dom/DeviceOrientation/window-property-expected.txt

Tests that the window.DeviceOrientationEvent and window.ondeviceorientation properties are present.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS typeof window.DeviceOrientationEvent == 'object' is true
PASS typeof window.DeviceOrientationEvent == 'function' is false
PASS hasDeviceOrientationEventProperty() is true
PASS 'DeviceOrientationEvent' in window is true
PASS window.hasOwnProperty('DeviceOrientationEvent') is true
PASS typeof window.ondeviceorientation == 'object' is true
PASS hasOnDeviceOrientationProperty() is true
PASS 'ondeviceorientation' in window is true
PASS window.hasOwnProperty('ondeviceorientation') is true
PASS successfullyParsed is true

TEST COMPLETE
